feat/network_utils
utils to check for network connection
